ml/note-service/requirements.txt
fastapi
google-generativeai
uvicorn[standard]
celery
pika
pydantic
pydantic-settings
boto3
transformers==4.44.2